Resolve DiagramWatcher terminal spam using a simple flag (#30)
Created a new object, the DiagramWatcherOptions, which will let users turn off some of the terminal messages during the DiagramWatcher's normal operation.

Also, modified all of the examples so that if one throws an Exception, the entire example fails. This should catch broken examples (at least broken in the sense that they throw an exception).